Made it compile on linux.
git-svn-id: https://wiki.csc.calpoly.edu/svn/AsteroidBlaster@511 342171fe-4ba3-422f-b357-09b6f24955fe

+++ Graphics/Skybox.cpp
@@ -25,7 +25,7 @@ void Skybox::draw(Camera* camera) {
    const bool rotateSkybox = false;
    if (rotateSkybox) {
       const double period = 60; // seconds
-      glRotatef(fmod(doubleTime(), period) * (360/ period), 1, 1, 1);
+      glRotatef( (GLfloat)(fmod(doubleTime(), period) * (360/ period)), 1, 1, 1);
    }
    glScalef(3, 3, 3);
    drawcube();
